feat: add export utilities
